Move RepositoriesClient to lib/
